fake_video_detector/requirements.txt
streamlit==1.22.0
requests==2.28.2
python-dotenv==1.0.0
validators==0.20.0 